Fix "city" id variable in Modifier class being int instead of long

+++ src/main/java/com/watersfall/clocgame/model/modifier/Modifier.java
@@ -10,7 +10,7 @@ public @AllArgsConstructor class Modifier
 {
 	private @Getter long id;
 	private @Getter int user;
-	private @Getter int city;
+	private @Getter long city;
 	private @Getter Modifiers type;
 	private @Getter long start;
 
@@ -18,7 +18,7 @@ public @AllArgsConstructor class Modifier
 	{
 		this.id = results.getLong("id");
 		this.user = results.getInt("user");
-		this.city = results.getInt("city");
+		this.city = results.getLong("city");
 		this.type = Modifiers.valueOf(results.getString("type"));
 		this.start = results.getLong("start");
 	}
